+++ packages/core/src/types/unified.ts
@@ -1,0 +1,209 @@
+import {
+  Affiliate,
+  AffiliateJSON,
+  Asset,
+  AssetJSON,
+  MultiChainMsg,
+  MultiChainMsgJSON,
+  PostHandler,
+  PostHandlerJSON,
+  Reason,
+  Swap,
+  SwapJSON,
+  SwapVenue,
+  SwapVenueJSON,
+  Transfer,
+  TransferJSON,
+} from "./shared";
+
+export type AssetsRequestJSON = {
+  chain_id?: string;
+  native_only?: boolean;
+  include_no_metadata_assets?: boolean;
+  include_cw20_assets?: boolean;
+};
+
+export type AssetsRequest = {
+  chainID?: string;
+  nativeOnly?: boolean;
+  includeNoMetadataAssets?: boolean;
+  includeCW20Assets?: boolean;
+};
+
+export type AssetsFromSourceRequestJSON = {
+  source_asset_denom: string;
+  source_asset_chain_id: string;
+
+  allow_multi_tx: boolean | undefined;
+  recommendation_reason: Reason | undefined;
+  include_swaps: boolean | undefined;
+  swap_venues: SwapVenueJSON[] | undefined;
+  native_only: boolean | undefined;
+  group_by: string | undefined;
+  include_cw20_assets: boolean;
+};
+
+export type AssetsFromSourceRequest = {
+  sourceAssetDenom: string;
+  sourceAssetChainID: string;
+
+  allowMultiTx: boolean | undefined;
+  recommendationReason: Reason | undefined;
+  includeSwaps: boolean | undefined;
+  swapVenues: SwapVenue[] | undefined;
+  nativeOnly: boolean | undefined;
+  groupBy: string | undefined;
+  includeCW20Assets: boolean;
+};
+
+export type AssetRecommendation = {
+  asset: Asset;
+  reason: Reason;
+};
+
+export type AssetRecommendationJSON = {
+  asset: AssetJSON;
+  reason: Reason;
+};
+
+export type RecommendAssetsRequestJSON = {
+  source_asset_denom: string;
+  source_asset_chain_id: string;
+  dest_chain_id: string;
+  reason?: Reason;
+};
+
+export type RecommendAssetsRequest = {
+  sourceAssetDenom: string;
+  sourceAssetChainID: string;
+  destChainID: string;
+  reason?: Reason;
+};
+
+export type RouteRequestBaseJSON = {
+  source_asset_denom: string;
+  source_asset_chain_id: string;
+  dest_asset_denom: string;
+  dest_asset_chain_id: string;
+
+  cumulative_affiliate_fee_bps?: string;
+  swap_venue?: SwapVenueJSON;
+};
+
+export type RouteRequestGivenInJSON = RouteRequestBaseJSON & {
+  amount_in: string;
+  amount_out?: never;
+};
+
+export type RouteRequestGivenOutJSON = RouteRequestBaseJSON & {
+  amount_in?: never;
+  amount_out: string;
+};
+
+export type RouteRequestJSON =
+  | RouteRequestGivenInJSON
+  | RouteRequestGivenOutJSON;
+
+export type RouteRequestBase = {
+  sourceAssetDenom: string;
+  sourceAssetChainID: string;
+  destAssetDenom: string;
+  destAssetChainID: string;
+
+  cumulativeAffiliateFeeBPS?: string;
+  swapVenue?: SwapVenue;
+};
+
+export type RouteRequestGivenIn = RouteRequestBase & {
+  amountIn: string;
+  amountOut?: never;
+};
+
+export type RouteRequestGivenOut = RouteRequestBase & {
+  amountIn?: never;
+  amountOut: string;
+};
+
+export type RouteRequest = RouteRequestGivenIn | RouteRequestGivenOut;
+
+export type OperationJSON = { transfer: TransferJSON } | { swap: SwapJSON };
+
+export type Operation = { transfer: Transfer } | { swap: Swap };
+
+export type RouteResponseJSON = {
+  source_asset_denom: string;
+  source_asset_chain_id: string;
+  dest_asset_denom: string;
+  dest_asset_chain_id: string;
+  amount_in: string;
+  amount_out: string;
+
+  operations: OperationJSON[];
+  chain_ids: string[];
+
+  does_swap: boolean;
+  estimated_amount_out?: string;
+  swap_venue?: SwapVenueJSON;
+
+  txs_required: number;
+};
+
+export type RouteResponse = {
+  sourceAssetDenom: string;
+  sourceAssetChainID: string;
+  destAssetDenom: string;
+  destAssetChainID: string;
+  amountIn: string;
+  amountOut: string;
+
+  operations: Operation[];
+  chainIDs: string[];
+
+  doesSwap: boolean;
+  estimatedAmountOut?: string;
+  swapVenue?: SwapVenue;
+
+  txsRequired: number;
+};
+
+export type MsgsRequestJSON = {
+  source_asset_denom: string;
+  source_asset_chain_id: string;
+  dest_asset_denom: string;
+  dest_asset_chain_id: string;
+  amount_in: string;
+  amount_out: string;
+  address_list: string[];
+  operations: OperationJSON[];
+
+  estimated_amount_out?: string;
+  slippage_tolerance_percent?: string;
+  affiliates?: AffiliateJSON[];
+
+  post_route_handler?: PostHandlerJSON;
+};
+
+export type MsgsRequest = {
+  sourceAssetDenom: string;
+  sourceAssetChainID: string;
+  destAssetDenom: string;
+  destAssetChainID: string;
+  amountIn: string;
+  amountOut: string;
+  addressList: string[];
+  operations: Operation[];
+
+  estimatedAmountOut?: string;
+  slippageTolerancePercent?: string;
+  affiliates?: Affiliate[];
+
+  postRouteHandler?: PostHandler;
+};
+
+export type MsgsResponseJSON = {
+  msgs: MultiChainMsgJSON[];
+};
+
+export type MsgsResponse = {
+  msgs: MultiChainMsg[];
+};
